+++ lab4/Customer.cpp
@@ -55,7 +55,7 @@ void Customer::displayHistory() const
 {
 	cout << "Customer ID = " << id << " " << lastName << " " << firstName;
 	
-	for (int i = 0; i < transHistory.size(); i++)
+	for (int i = 0; i < (int)transHistory.size(); i++)
 	{
 		cout << " ";
 		transHistory[i].Display();
